Parse command line before App launch (#746)
